routed static files and templates
Had to change the api => app for reference in the app file, and change api.py to app.py
