--- api.py
@@ -1,7 +1,0 @@
-from flask import Flask
-
-api = Flask(__name__)
-
-@api.route('/')
-def hello_world():
-    return ('Hello World from the backend')
