+++ static/css/home.css
@@ -1,0 +1,27 @@
+﻿body {
+    margin: 0;
+    padding: 0;
+    background: #ecf1f1;
+    color: black;
+}
+
+footer {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: fixed;
+    width: 100vw;
+    bottom: 0px;
+    color: inherit;
+    border-top: 4px double #997555;
+    background: linear-gradient(45deg, transparent, #f0ecef, transparent);
+}
+
+
+#hero_image {
+    width: 5vw;
+    height: 10vh;
+}
+
+/* light back: #ecf1f1*/
+/* dark back: #25262d*/
