+++ templates/SAS_index.html
@@ -8,11 +8,11 @@
     <meta name="author" content="Regi & Nellie">
     <meta name="description" content="Created for Chingu Voyage 22, a site to discover the possibilities of accessibility.">
     <!-- JavaScript -->
-    <script type="text/javascript" src="ham_nav.js"></script>
+    <script type="text/javascript" src={{ url_for('static', filename='js/ham_nav.js')}}></script>
     <!-- Style sheets -->
-    <link rel="stylesheet" href="home.css" />
-    <link rel="stylesheet" href="ham_nav.css" />
-    <link rel="icon" href="SAS_logo.png"/>
+    <link rel="stylesheet" href={{ url_for('static', filename='css/home.css')}} />
+    <link rel="stylesheet" href={{ url_for('static', filename='css/ham_nav.css')}} />
+    <link rel="icon" href={{ url_for('static', filename='images/SAS_log.png')}}/>
     <!-- Font awesome SVG-->
     <script src="https://kit.fontawesome.com/b570a31fd0.js" crossorigin="anonymous"></script>
 
@@ -31,7 +31,7 @@
 
 </main>
 <footer id="footer">
-    <image src="SAS_logo.png" id="hero_image" alt="A sillouette of a head with gaming headphones on."></image>
+    <image src={{ url_for('static', filename='images/SAS_logo.png')}} id="hero_image" alt="A sillouette of a head with gaming headphones on."></image>
     <h3 class="footer-info">Created by Chingu Bear team:  Regi and Nellie</h3>
 </footer>
 </body>
